Move `get_aggregation_time_dimensions_with_measures` to `MeasureLookup`
