+++ metricflow-semantics/metricflow_semantics/model/semantics/semantic_model_lookup.py
@@ -19,14 +19,11 @@ from dbt_semantic_interfaces.type_enums import DimensionType
 from metricflow_semantics.errors.error_classes import InvalidSemanticModelError
 from metricflow_semantics.mf_logging.lazy_formattable import LazyFormat
 from metricflow_semantics.model.semantics.dimension_lookup import DimensionLookup
-from metricflow_semantics.model.semantics.element_group import ElementGrouper
 from metricflow_semantics.model.semantics.measure_lookup import MeasureLookup
-from metricflow_semantics.model.spec_converters import MeasureConverter
 from metricflow_semantics.naming.linkable_spec_name import StructuredLinkableSpecName
 from metricflow_semantics.specs.dimension_spec import DimensionSpec
 from metricflow_semantics.specs.entity_spec import EntitySpec
 from metricflow_semantics.specs.instance_spec import LinkableInstanceSpec
-from metricflow_semantics.specs.measure_spec import MeasureSpec
 from metricflow_semantics.specs.time_dimension_spec import DEFAULT_TIME_GRANULARITY, TimeDimensionSpec
 from metricflow_semantics.time.granularity import ExpandedTimeGranularity
 
@@ -48,10 +45,6 @@ class SemanticModelLookup:
 
         self._dimension_ref_to_spec: Dict[DimensionReference, DimensionSpec] = {}
         self._entity_ref_to_spec: Dict[EntityReference, EntitySpec] = {}
-
-        self._semantic_model_to_aggregation_time_dimensions: Dict[
-            SemanticModelReference, ElementGrouper[TimeDimensionReference, MeasureSpec]
-        ] = {}
 
         self._semantic_model_reference_to_semantic_model: Dict[SemanticModelReference, SemanticModel] = {}
         sorted_semantic_models = sorted(model.semantic_models, key=lambda semantic_model: semantic_model.name)
@@ -98,41 +91,10 @@ class SemanticModelLookup:
             errors.append(f"Semantic model {semantic_model.reference} already added.")
 
         if len(errors) > 0:
-            raise InvalidSemanticModelError(f"Error adding {semantic_model.reference}. Got errors: {errors}")
-
-        self._semantic_model_to_aggregation_time_dimensions[semantic_model.reference] = ElementGrouper[
-            TimeDimensionReference, MeasureSpec
-        ]()
-
-        for measure in semantic_model.measures:
-            agg_time_dimension_reference = semantic_model.checked_agg_time_dimension_for_measure(measure.reference)
-            # Ensure agg_time_dimension is lowercased - this transformation was not enforced on earlier manifests
-            agg_time_dimension_reference = TimeDimensionReference(agg_time_dimension_reference.element_name.lower())
-
-            matching_dimensions = tuple(
-                dimension
-                for dimension in semantic_model.dimensions
-                if dimension.type == DimensionType.TIME
-                and dimension.time_dimension_reference == agg_time_dimension_reference
-            )
-
-            matching_dimensions_count = len(matching_dimensions)
-            if matching_dimensions_count != 1:
-                raise RuntimeError(
-                    f"Found {matching_dimensions_count} matching dimensions for {agg_time_dimension_reference} "
-                    f"in {semantic_model}"
+            raise InvalidSemanticModelError(
+                LazyFormat(
+                    "Got errors adding the given semantic model.", semantic_model=semantic_model.name, errors=errors
                 )
-            agg_time_dimension = matching_dimensions[0]
-            if agg_time_dimension.type_params is None or agg_time_dimension.type_params.time_granularity is None:
-                raise RuntimeError(
-                    f"Aggregation time dimension does not have a time granularity set: {agg_time_dimension}"
-                )
-
-            self._semantic_model_to_aggregation_time_dimensions[semantic_model.reference].add_value(
-                key=TimeDimensionReference(
-                    element_name=agg_time_dimension.name,
-                ),
-                value=MeasureConverter.convert_to_measure_spec(measure=measure),
             )
 
         for dim in semantic_model.dimensions:
@@ -168,15 +130,6 @@ class SemanticModelLookup:
             self._entity_ref_to_spec[entity.reference] = EntitySpec(element_name=entity.name, entity_links=())
 
         self._semantic_model_reference_to_semantic_model[semantic_model.reference] = semantic_model
-
-    def get_aggregation_time_dimensions_with_measures(
-        self, semantic_model_reference: SemanticModelReference
-    ) -> ElementGrouper[TimeDimensionReference, MeasureSpec]:
-        """Return all time dimensions in a semantic model with their associated measures."""
-        assert (
-            semantic_model_reference in self._semantic_model_to_aggregation_time_dimensions
-        ), f"Semantic Model {semantic_model_reference} is not known"
-        return self._semantic_model_to_aggregation_time_dimensions[semantic_model_reference]
 
     def get_semantic_models_for_entity(self, entity_reference: EntityReference) -> Set[SemanticModel]:
         """Return all semantic models associated with an entity reference."""
