+++ metricflow/dataflow/builder/source_node.py
@@ -90,10 +90,9 @@ class SourceNodeBuilder:
         for data_set in data_sets:
             read_node = ReadSqlSourceNode.create(data_set)
             group_by_item_source_nodes.append(read_node)
-            agg_time_dim_to_measures_grouper = (
-                self._semantic_manifest_lookup.semantic_model_lookup.get_aggregation_time_dimensions_with_measures(
-                    data_set.semantic_model_reference
-                )
+            measure_lookup = self._semantic_manifest_lookup.semantic_model_lookup.measure_lookup
+            agg_time_dim_to_measures_grouper = measure_lookup.get_aggregation_time_dimensions_with_measures(
+                data_set.semantic_model_reference
             )
 
             # Dimension sources may not have any measures -> no aggregation time dimensions.
